+++ lib/generators/templates/easy_captcha.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 EasyCaptcha.setup do |config|
-  # ### Cache
+  # #####
+  # # Cache
   # config.cache                   = true
   # # Cache temp dir from Rails.root
   # config.cache_temp_dir          = Rails.root + 'tmp' + 'captchas'
@@ -10,7 +11,8 @@ EasyCaptcha.setup do |config|
   # # Cache max length of time to cache an image
   # config.cache_expire            = 1.days
   #
-  # ### CAPTCHA
+  # #####
+  # # CAPTCHA
   # # Chars available for CAPTCHA
   # config.captcha_character_pool  = %w(2 3 4 5 6 7 9 A C D E F G H J K L M N P Q R S T U X Y Z)
   #
@@ -21,7 +23,8 @@ EasyCaptcha.setup do |config|
   # config.captcha_image_height    = 40
   # config.captcha_image_width     = 140
 
-  # ### eSpeak
+  # #####
+  # # eSpeak
   # # Enable eSpeak using all defaults:
   # config.espeak = true
   #
@@ -40,7 +43,8 @@ EasyCaptcha.setup do |config|
   #   espeak.voice = nil
   # end
 
-  # ### Generator
+  # #####
+  # # Generator
   # config.generator :default do |generator|
   #   # Blur
   #   generator.blur                   = true
@@ -48,26 +52,26 @@ EasyCaptcha.setup do |config|
   #   generator.blur_sigma             = 2
   #
   #   # Font
-  #   generator.font_size              = 24
-  #   generator.font_fill_color        = '#333333'
-  #   generator.font_stroke_color      = '#000000'
-  #   generator.font_stroke            = 0
   #   generator.font_family            = File.expand_path('../../resources/afont.ttf', __FILE__)
+  #   generator.font_fill_color        = '#333333'
+  #   generator.font_size              = 24
+  #   generator.font_stroke            = 0
+  #   generator.font_stroke_color      = '#000000'
   #
+  #   # Image
   #   generator.image_background_color = "#FFFFFF"
   #
-  #   # Wave
-  #   generator.wave                   = true
-  #   generator.wave_length            = (60..100)
-  #   generator.wave_amplitude         = (3..5)
+  #   # Implode
+  #   generator.implode                = 0.1
   #
   #   # Sketch
   #   generator.sketch                 = true
   #   generator.sketch_radius          = 3
   #   generator.sketch_sigma           = 1
   #
-  #   # Implode
-  #   generator.implode                = 0.1
-  #
+  #   # Wave
+  #   generator.wave                   = true
+  #   generator.wave_amplitude         = (3..5)
+  #   generator.wave_length            = (60..100)
   # end
 end