+++ CHANGELOG.md
@@ -3,12 +3,38 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.9.0 (TBD)
+## 0.9.0 (2020-12-05)
+
+### Additions
+
+* [karlwilbur] - Add tests for:
+
+    - `EasyCaptcha::Captcha`
+    - `EasyCaptcha::CaptchaController`
+    - `EasyCaptcha::ControllerHelpers`
+    - `EasyCaptcha::ModelHelpers`
+    - `EasyCaptcha::ViewHelpers`
+    - `EasyCaptcha::Espeak` WAV file generation
+    - `EasyCaptcha::Generators::InstallGenerator`
 
 * [karlwilbur] - Add `mdl` gem dev dependency; run linters using `bundle exec`
 * [karlwilbur] - Add `CHANGELOG`
 * [karlwilbur] - Add CodeClimate badges to `README`
+
+### Modifications/Fixes
+
+* [karlwilbur] - Refactor `EasyCaptcha::Espeak` to use `DEFAULT_CONFIG` constant
+* [karlwilbur] - Reduce complexity of `EasyCaptcha::Espeak#generate`
+* [karlwilbur] - Reduce complexity of `EasyCaptcha#generator`
+* [karlwilbur] - Fix logic in `EasyCaptcha::ModelHelper#captcha_valid?`
+* [karlwilbur] - Fix bug in audio CAPTCHA generation
+* [karlwilbur] - Modify comments and order of config options in initializer template
+* [karlwilbur] - Allow `EasyCaptcha#setup` to be called without a block to use defaults
 * [karlwilbur] - Lock Simplecov version until CodeClimate Test Reporter is updated
+
+### Deletions/Removals
+
+* [karlwilbur] - Remove deprecation warnings for deprecated methods which were already removed
 
 ## 0.8.0 (2020-12-03)